refactor: improve error handling in environment middleware and update imports
