+++ src/middleware/envExists.ts
@@ -4,13 +4,14 @@ export const envExists = (req: Request, res: Response, next: NextFunction) => {
   const { name } = req.params;
 
   if (!name) {
-    return res.status(400).json({ error: 'Environment name is required' });
+    res.status(400).json({ error: 'Environment name is required' });
+    return;
   }
 
-  // Placeholder: Check DynamoDB later
   const exists = true;
   if (!exists) {
-    return res.status(404).json({ error: `Environment ${name} not found` });
+    res.status(404).json({ error: `Environment ${name} not found` });
+    return;
   }
 
   next();
