+++ src/middleware/validateEnvRequest.ts
@@ -1,17 +1,20 @@
 import { Request, Response, NextFunction } from 'express';
-import { EnvironmentRequest } from '../types/envTypes'; // Updated import
+import { EnvironmentRequest } from '../types/envTypes';
 
-export const validateEnvRequest = (req: Request, res: Response, next: NextFunction) => {
+export const validateEnvRequest = (req: Request, res: Response, next: NextFunction): void => {
   const { name, image, ttl } = req.body as Partial<EnvironmentRequest>;
 
   if (!name || typeof name !== 'string' || name.length < 3) {
-    return res.status(400).json({ error: 'Invalid or missing environment name (min 3 chars)' });
+    res.status(400).json({ error: 'Invalid or missing environment name (min 3 chars)' });
+    return;
   }
   if (!image || typeof image !== 'string') {
-    return res.status(400).json({ error: 'Invalid or missing image' });
+    res.status(400).json({ error: 'Invalid or missing image' });
+    return;
   }
   if (!ttl || typeof ttl !== 'number' || ttl <= 0) {
-    return res.status(400).json({ error: 'Invalid or missing TTL (must be positive number)' });
+    res.status(400).json({ error: 'Invalid or missing TTL (must be positive number)' });
+    return;
   }
 
   next();
