+++ src/routes/envRoutes.ts
@@ -1,6 +1,7 @@
 import { Router } from 'express';
 import { createEnvironment, getEnvironment } from '../controllers/envController';
-import { validateEnvRequest, envExists } from '../middleware';
+import { validateEnvRequest } from '../middleware/validateEnvRequest';
+import { envExists } from '../middleware/envExists';
 
 const router = Router();
